+++ lib/src/rules/avoid_preserve_whitespace_false.dart
@@ -1,0 +1,56 @@
+import 'package:analyzer/dart/ast/ast.dart';
+import 'package:analyzer/dart/ast/token.dart';
+import 'package:analyzer/dart/ast/visitor.dart';
+import 'package:dart_code_metrics/src/models/code_issue.dart';
+import 'package:dart_code_metrics/src/models/code_issue_severity.dart';
+
+import 'base_rule.dart';
+import 'rule_utils.dart';
+
+class AvoidPreserveWhitespaceFalseRule extends BaseRule {
+  static const _failure = 'Avoid using preserveWhitespace: false.';
+
+  const AvoidPreserveWhitespaceFalseRule()
+      : super(
+          id: 'avoid-preserve-whitespace-false',
+          severity: CodeIssueSeverity.warning,
+        );
+
+  @override
+  Iterable<CodeIssue> check(
+      CompilationUnit unit, Uri sourceUrl, String sourceContent) {
+    final visitor = _Visitor();
+
+    unit.visitChildren(visitor);
+
+    return visitor.expression
+        .map((expression) => createIssue(this, _failure, null, null, sourceUrl,
+            sourceContent, unit.lineInfo, expression))
+        .toList();
+  }
+}
+
+class _Visitor extends RecursiveAstVisitor<Object> {
+  final _expression = <NamedExpression>[];
+
+  Iterable<NamedExpression> get expression => _expression;
+
+  @override
+  void visitAnnotation(Annotation node) {
+    if (node.name.name == 'Component' &&
+        node.atSign.type.lexeme == '@' &&
+        node.parent is ClassDeclaration) {
+      final preserveWhitespaceArg = node.arguments.arguments
+          .whereType<NamedExpression>()
+          .firstWhere((arg) => arg.name.label.name == 'preserveWhitespace',
+              orElse: () => null);
+      if (preserveWhitespaceArg != null) {
+        final expression = preserveWhitespaceArg.expression;
+        if (expression is BooleanLiteral &&
+            expression.literal.keyword == Keyword.FALSE) {
+          _expression.add(preserveWhitespaceArg);
+        }
+      }
+    }
+  }
+}