+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 # Unreleased
 - Improve code diagnostics double-literal-format, no-boolean-literal-compare
-- Add experimental static code diagnostics newline-before-return, no-empty-block
+- Add experimental static code diagnostics newline-before-return, no-empty-block, avoid-preserve-whitespace-false
 - Support cyclomatic complexity metric in analyzer plugin
 
 # 1.5.0